+++ pkg/sdk/shares_integration_test.go
@@ -118,7 +118,7 @@ func TestInt_SharesDrop(t *testing.T) {
 
 	t.Run("when share does not exist", func(t *testing.T) {
 		err := client.Shares.Drop(ctx, NewAccountObjectIdentifier("does_not_exist"))
-		assert.ErrorIs(t, err, ErrObjectNotExistOrAuthorized)
+		assert.ErrorIs(t, err, errObjectNotExistOrAuthorized)
 	})
 }