+++ pkg/sdk/warehouses.go
@@ -125,7 +125,7 @@ type CreateWarehouseOptions struct {
 
 func (opts *CreateWarehouseOptions) validate() error {
 	if !validObjectidentifier(opts.name) {
-		return ErrInvalidObjectIdentifier
+		return errInvalidObjectIdentifier
 	}
 	if valueSet(opts.MinClusterCount) && valueSet(opts.MaxClusterCount) && !validateIntGreaterThanOrEqual(*opts.MaxClusterCount, *opts.MinClusterCount) {
 		return fmt.Errorf("MinClusterCount must be less than or equal to MaxClusterCount")
@@ -170,7 +170,7 @@ type AlterWarehouseOptions struct {
 
 func (opts *AlterWarehouseOptions) validate() error {
 	if !validObjectidentifier(opts.name) {
-		return ErrInvalidObjectIdentifier
+		return errInvalidObjectIdentifier
 	}
 	if ok := exactlyOneValueSet(
 		opts.Suspend,
@@ -302,7 +302,7 @@ type DropWarehouseOptions struct {
 
 func (opts *DropWarehouseOptions) validate() error {
 	if !validObjectidentifier(opts.name) {
-		return ErrInvalidObjectIdentifier
+		return errInvalidObjectIdentifier
 	}
 	return nil
 }
@@ -492,7 +492,7 @@ func (c *warehouses) ShowByID(ctx context.Context, id AccountObjectIdentifier) (
 			return warehouse, nil
 		}
 	}
-	return nil, ErrObjectNotExistOrAuthorized
+	return nil, errObjectNotExistOrAuthorized
 }
 
 type warehouseDescribeOptions struct {
@@ -503,7 +503,7 @@ type warehouseDescribeOptions struct {
 
 func (opts *warehouseDescribeOptions) validate() error {
 	if !validObjectidentifier(opts.name) {
-		return ErrInvalidObjectIdentifier
+		return errInvalidObjectIdentifier
 	}
 	return nil
 }